--- src/components/campaign/CampaignForm.scss
@@ -1,132 +1,0 @@
-.CampaignForm {
-    position: relative;
-    z-index: 1;
-    background: darken(white, 1.5);
-    width: 100%;
-    float:left;
-}
-
-.CampaignForm-calendar {
-    @include col(10,12);
-    @include col-offset(1, 12);
-}
-
-.CampaignForm-filter {
-    @include col(11,12);
-    @include col-offset(1, 12);
-    margin-top: 1em;
-}
-
-.CampaignForm-form {
-    @include col(12,12);
-}
-
-.CampaignForm-actions {
-    @include col(12,12);
-    .CampaignForm-action {
-        position: relative;
-        margin: 1em 0;
-        list-style: none;
-
-        &:before {
-            content: " ";
-            background: darken(white, 10);
-            height: 1.5em;
-            width: 1.5em;
-            display: block;
-            border-radius: 50%;
-            position: absolute;
-            margin-left: -2.6em;
-            top: calc(50% - 0.5em);
-            box-shadow: 0 2px 3px rgba(0, 0, 0, 0.15);
-        }
-        &.response:before {
-            content: "●";
-            text-align: center;
-            line-height: 1.5em;
-            color: $c-brand-main;
-        }
-        &.booked:before {
-            background: $c-brand-main;
-        }
-
-        > div {
-            @include card;
-            padding: 1em;
-            position: relative;
-
-            ul {
-                padding: 0;
-                list-style: none;
-            }
-        }
-    }
-}
-.CampaignForm-days {
-    list-style: none;
-    padding: 0;
-    margin-left: 2.5em;
-    margin-top: 2em;
-
-    &:before {
-        content: " ";
-        background: darken(white, 10);
-        height: 100%;
-        width: 0.2em;
-        display: block;
-        position: absolute;
-        top: 0;
-        margin-left: -1.45em;
-    }
-
-    .CampaignForm-day {
-        margin-bottom: 1em;
-        .CampaignForm-date {
-            color: $c-text;
-            position: absolute;
-            margin-left: 1em;
-            margin-top: -1em;
-            text-align: right;
-            font-size: 0.8em;
-        }
-    }
-}
-
-@include medium-screen {
-    .Dashboard {
-        & > div {
-            float: left;
-            @include col(6, 12);
-        }
-    }
-}
-
-@include large-screen {
-    .CampaignForm-calendar {
-        @include col(3,12);
-        @include col-offset(0, 12);
-        padding-left: 2em;
-        float: left;
-        position: absolute;
-        left: 0;
-    }
-
-    .CampaignForm-filter {
-        @include col(8,12);
-        @include col-offset(4, 12);
-    }
-
-    .CampaignForm-form {
-        @include col(5,12);
-        @include col-offset(3, 12);
-    }
-    .CampaignForm-days {
-        margin-left: 6em;
-
-        .CampaignForm-day {
-            .CampaignForm-date {
-                margin-left: -6em;
-            }
-        }
-    }
-}
