--- src/components/campaign/action/ActionFormLocation.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-
-export default function ActionFormLocation(props) {
-    return (
-        <div className="ActionFormLocation">
-            { props.location }
-        </div>
-    );
-};
