--- src/components/campaign/action/ActionFormLocation.scss
@@ -1,9 +1,0 @@
-.ActionFormLocation {
-    font-size: 0.9em;
-    color: lighten($c-text, 20);
-    margin-bottom: 0.2em;
-
-    &:before {
-        @include icon($fa-var-map-marker);
-    }
-}
