--- src/components/campaign/action/ActionFormTime.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-
-export default function ActionFormTime(props) {
-    return (
-        <div className="ActionFormTime">
-            { props.time }
-        </div>
-    );
-};
