--- src/components/campaign/action/ActionFormTitle.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-
-export default function ActionFormTitle(props) {
-    return (
-        <h3 className="ActionFormTitle">
-            { props.title }
-        </h3>
-    );
-};
