--- src/components/campaign/action/ActionFormTitle.scss
@@ -1,4 +1,0 @@
-.ActionFormTitle {
-    margin: 0.5em 0 0.2em;
-    font-size: 1.2em;
-}
