--- src/components/campaign/action/MultiLocationActionForm.scss
@@ -1,22 +1,0 @@
-.MultiLocationActionForm {
-
-    .MultiLocationActionForm-locationsLink {
-        text-decoration: underline;
-        cursor: pointer;
-        color: lighten($c-text, 20);
-
-        &:before {
-            @include icon($fa-var-map-marker);
-        }
-    }
-
-    .MultiLocationActionForm-locationItem {
-        position: relative;
-        margin: 0.5em 0;
-
-        .ResponseWidget {
-            right: 0;
-            bottom: 0;
-        }
-    }
-}
