--- src/components/campaign/action/MultiShiftActionForm.scss
@@ -1,9 +1,0 @@
-.MultiShiftActionForm-shiftItem {
-    position: relative;
-    margin: 0.5em 0;
-
-    .ResponseWidget {
-        right: 0;
-        bottom: 0;
-    }
-}
