--- src/components/campaign/action/SingleActionForm.scss
@@ -1,59 +1,0 @@
-.SingleActionForm-title {
-    margin: 0.5em 0 0.2em;
-    font-size: 1.2em;
-}
-.SingleActionForm-location, .SingleActionForm-time {
-    font-size: 0.9em;
-    color: lighten($c-text, 20);
-    margin-bottom: 0.2em;
-}
-.SingleActionForm-location {
-    &:before {
-        @include icon($fa-var-map-marker);
-    }
-}
-.SingleActionForm-time {
-    &:before {
-        @include icon($fa-var-clock-o);
-    }
-}
-.SingleActionForm-info {
-    color: lighten($c-text, 30);
-    font-size: 0.8em;
-    line-height: 1.5em;
-    margin: 0 3.2em 0 0;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    max-height: 3em;
-    transition: max-height 0.5s;
-}
-.SingleActionForm-toggleExpandButton {
-    bottom: 0;
-    margin-top: 0.5em;
-    margin-left: calc(50% - 1em);
-}
-.SingleActionForm.contracted {
-    &:after {
-        content: "";
-        position: absolute;
-        right: 0;
-        bottom: 3em;
-        width: 40%;
-        height: 1.5em;
-        background: linear-gradient(to right, 
-            rgba(255, 255, 255, 0), 
-            rgba(255, 255, 255, 1) 50%);
-    }
-    .SingleActionForm-toggleExpandButton {
-        @include circle-button($icon:$fa-var-caret-down);
-    }
-}
-.SingleActionForm.expanded {
-    .SingleActionForm-info {
-        max-height: 10em;
-    }
-
-    .SingleActionForm-toggleExpandButton {
-        @include circle-button($icon:$fa-var-caret-up);
-    }
-}
