--- src/components/campaign/calendar/CampaignCalendar.scss
@@ -1,22 +1,0 @@
-.CampaignCalendar {
-    margin-top: 2em;
-
-    &:after {
-        content: "";
-        display: block;
-        clear: both;
-    }
-}
-.CampaignCalendar-header {
-    ul {
-        list-style: none;
-        padding-left: 8%;
-        margin: 0 0 0.5em 0;
-    }
-    li {
-        display: inline-block;
-        width: 14%;
-        text-align: center;
-        color: lighten($c-text, 65);
-    }
-}
