--- src/components/campaign/calendar/CampaignCalendarWeek.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-
-
-export default class CampaignCalendarWeek extends React.Component {
-    render() {
-        return (
-            <div className="CampaignCalendarWeek">
-                <span className="CampaignCalendarWeek-number">
-                    { this.props.week }
-                </span>
-                <ul className="CampaignCalendarWeek-days">
-                    { this.props.children }
-                </ul>
-            </div>
-        );
-    }
-}
