--- src/components/campaign/calendar/CampaignCalendarWeek.scss
@@ -1,23 +1,0 @@
-.CampaignCalendarWeek {
-    clear: left;
-    width: 100%;
-}
-
-.CampaignCalendarWeek-number {
-    float: left;
-    font-size: 0.8em;
-    line-height: 3em;
-    color: lighten($c-text, 30);
-    width: 8%;
-    height: 0;
-    padding-bottom: 13%;
-}
-
-.CampaignCalendarWeek-days {
-    float: left;
-    margin: 0;
-    padding: 0;
-    border-bottom: 1px solid darken(white, 5);
-    border-left: 1px solid darken(white, 5);
-    width: 92%;
-}
