--- src/components/campaign/filter/CampaignFilter.scss
@@ -1,3 +1,0 @@
-.CampaignFilter h3 {
-    display: none;
-}
