--- src/components/campaign/filter/CampaignFilterList.scss
@@ -1,62 +1,0 @@
-.CampaignFilterList {
-    padding: 0;
-    margin: 0.2em;
-    display: inline-block;
-
-    &:after {
-        content: "";
-        display: block;
-        clear: both;
-    }
-
-    &.contracted {
-        .CampaignFilterList-toggle {
-            &:after {
-                @include icon($fa-var-chevron-down);
-            }
-        }
-
-        .CampaignFilterList-list {
-            visibility: hidden;
-            max-height: 0;
-            max-width: 0;
-            overflow: hidden;
-            transition: all 0.5s;
-        }
-    }
-    &.expanded {
-        .CampaignFilterList-toggle {
-            &:after {
-                @include icon($fa-var-chevron-up);
-            }
-        }
-
-        .CampaignFilterList-list {
-            visibility: visible;
-            max-height: 100em;
-            overflow: hidden;
-            transition: all 0.5s;
-        }
-    }
-}
-
-.CampaignFilterList-list {
-    margin: 0;
-    padding-left: 0;
-}
-
-.CampaignFilterList-item {
-    list-style-type: none;
-    @include inline-button(white);
-    float: left;
-    font-size: 0.9em;
-    margin: 0.3em;
-}
-
-.CampaignFilterList-toggle {
-    //TODO: Update with new button mixin
-    @include button($color:darken(white, 5));
-    font-size: 0.8em;
-    padding: 0 1.5em;
-    cursor: pointer;
-}
