--- src/components/misc/LoadingIndicator.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-
-export default (props) => (
-    <span className="LoadingIndicator">
-        Loading...
-    </span>
-);
